+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 ## [v0.5.0](https://github.com/sikessem/ui/releases/tag/v0.5.0) - 2023-08-01
 
+- [0.x] Fix [color utilities](https://github.com/sikessem/ui#color-utilities)
 - [0.x] Add [color utilities](https://github.com/sikessem/ui#color-utilities)
 - [0.x] Add [icon](https://github.com/sikessem/ui#icon-component) component